create invocation handler only when needed
